a function that counts nodes

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t counter = 0;
+	unsigned int counter = 0;
 
 	while (h != NULL)
 	{
